+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # CHANGELOG
 
 ## Version 0.0.1
+- Added database schemas (#1)